Fix building on 64-bit platforms
git-svn-id: https://svn.code.sf.net/p/moonunit/svn/trunk/moonunit@127 7cfde7b9-103a-0410-917d-a4b0b378e324

+++ include/uipc/marshal.h
@@ -40,7 +40,7 @@ typedef struct __uipc_typeinfo
 #define UIPC_OFFSET(type, field) ((unsigned long) &((type*)0)->field)
 #define UIPC_POINTER(type, field, info) {UIPC_OFFSET(type, field), info}
 
-void uipc_marshal_payload(void* membase, unsigned int memsize, void* payload, uipc_typeinfo* payload_type);
-void uipc_unmarshal_payload(void* membase, unsigned int memsize, void* payload, uipc_typeinfo* payload_type);
+void uipc_marshal_payload(void* membase, unsigned long memsize, void* payload, uipc_typeinfo* payload_type);
+void uipc_unmarshal_payload(void* membase, unsigned long memsize, void* payload, uipc_typeinfo* payload_type);
 
 #endif
